feat: + POST /bookings; + DELETE /bookings/{id}

+++ package/backend/internal/bookings/model.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Booking struct {
-	ID        int64     `json:"id"`
-	UserID    int64     `json:"user_id,omitempty"`
-	SlotID    int64     `json:"slot_id"`
+	ID        int       `json:"id"`
+	UserID    int       `json:"user_id,omitempty"`
+	SlotID    int       `json:"slot_id"`
 	CreatedAt time.Time `json:"created_at"`
 }
